made a helper function which makes an array of objects containing pokemon names and images. Merged it to the game cards object successfully.
We can now see the pokemon images, the names and the alt img are set up as well. I moved helper functions out of gameArea to clean up the code slightly.

game logic works, game loop works - all that is left is to make the tooltip and instructions areas, and then style.