I think it's a finished product. Finished up a lot of small touches for styling.
changed footer, to be falir and added it under the header.

made some changes to how the gameArea looks, and is spaced.

changes to score and best score.

overall css changes, and small look tweaks.

all that remains is writing the read me, and adding some media queries for larger screens.